+++ src/throwable.cpp
@@ -11,7 +11,7 @@ int Throwable::id_throwables = 0;
 Throwable::Throwable(Stage& stage
 	, const int x
 	, const int y
-	, const int angle_rad
+	, const float angle_rad
 	, const b2Vec2& velocity
 	, const float angular_velocity
 	, const float radius
