+++ src/worm.cpp
@@ -12,7 +12,7 @@ Worm::Worm(Stage& stage
 	, const int id
 	, const int x
 	, const int y
-	, const int angle_rad
+	, const float angle_rad
 	, const int longitude
 	, const int height
 	, const float restitution
